Lib reworked for a little bit better modularization
